Move todolist component to own folder
