+++ src/sentry/static/sentry/app/components/sidebar/index.jsx
@@ -12,7 +12,7 @@ import UserNav from './userNav';
 import requiredAdminActions from '../requiredAdminActions';
 import OrganizationSelector from './organizationSelector';
 import SidebarPanel from '../sidebarPanel';
-import TodoList from '../todos';
+import TodoList from '../onboardingWizard/todos';
 import IssueList from '../issueList';
 import ConfigStore from '../../stores/configStore';
